Raw Exports From an Example REDCap Project

These data sets are the results of calling export_content. An API token is required to reproduce these calls. No such token will be provided publicly, so these data sets are provided so end users can run examples for other tools provided in the REDCapExporter package.

Usage

avs_raw_project

avs_raw_metadata

avs_raw_user

avs_raw_record

avs_raw_core

avs_raw_project_json

avs_raw_metadata_json

avs_raw_user_json

avs_raw_record_json

avs_raw_core_json

Format

An object of class rcer_raw_project (inherits from character) of length 1.

An object of class rcer_raw_metadata (inherits from character) of length 1.

An object of class rcer_raw_user (inherits from character) of length 1.

An object of class rcer_raw_record (inherits from character) of length 1.

An object of class rcer_rccore of length 4.

An object of class rcer_raw_project (inherits from character) of length 1.

An object of class rcer_raw_metadata (inherits from character) of length 1.

An object of class rcer_raw_user (inherits from character) of length 1.

An object of class rcer_raw_record (inherits from character) of length 1.

An object of class rcer_rccore of length 4.

Details

avs_raw_project provides meta data about the project itself in a csv format. avs_raw_project_json is the same information in json format.

avs_raw_metadata is the data dictionary for the REDCap Project in a csv format. This information can be used with format_record to build a data.frame that is ready for analysis. avs_raw_metadata_json is the same metadata information in json format.

avs_raw_user REDCap Project user table in csv format. avs_raw_user_json is the same information in json format.

avs_raw_record REDCap Project records, i.e., 'the data' in csv format. avs_raw_record_json is the same information in json format.

Examples

if (FALSE) { # \dontrun{
avs_raw_project  <- export_content(content = "project",  format = "csv")
avs_raw_metadata <- export_content(content = "metadata", format = "csv")
avs_raw_user     <- export_content(content = "user",     format = "csv")
avs_raw_record   <- export_content(content = "record",   format = "csv")
avs_raw_core     <- export_core(format = "csv")
} # }

data(avs_raw_project)
data(avs_raw_user)
data(avs_raw_metadata)
data(avs_raw_record)
data(avs_raw_core)

str(avs_raw_project)
#>  'rcer_raw_project' chr "project_id,project_title,creation_time,production_time,in_production,project_language,purpose,purpose_other,pro"| __truncated__
#>  - attr(*, "url")= chr "https://redcap.ucdenver.edu/api/"
#>  - attr(*, "status_code")= int 200
#>  - attr(*, "times")= Named num [1:6] 0 0.0244 0.0371 0.0627 0.2281 ...
#>   ..- attr(*, "names")= chr [1:6] "redirect" "namelookup" "connect" "pretransfer" ...
#>  - attr(*, "Content-Type")= chr "text/csv; charset=utf-8"
#>  - attr(*, "accessed")= POSIXct[1:1], format: "2024-09-19 17:09:20"
str(avs_raw_user)
#>  'rcer_raw_user' chr "username,email,firstname,lastname,expiration,data_access_group,data_access_group_id,design,alerts,user_rights,d"| __truncated__
#>  - attr(*, "url")= chr "https://redcap.ucdenver.edu/api/"
#>  - attr(*, "status_code")= int 200
#>  - attr(*, "times")= Named num [1:6] 0 0.000013 0 0.000082 0.119041 ...
#>   ..- attr(*, "names")= chr [1:6] "redirect" "namelookup" "connect" "pretransfer" ...
#>  - attr(*, "Content-Type")= chr "text/csv; charset=utf-8"
#>  - attr(*, "accessed")= POSIXct[1:1], format: "2024-09-19 17:09:20"
str(avs_raw_metadata)
#>  'rcer_raw_metadata' chr "field_name,form_name,section_header,field_type,field_label,select_choices_or_calculations,field_note,text_valid"| __truncated__
#>  - attr(*, "url")= chr "https://redcap.ucdenver.edu/api/"
#>  - attr(*, "status_code")= int 200
#>  - attr(*, "times")= Named num [1:6] 0 0.000014 0 0.000076 0.131388 ...
#>   ..- attr(*, "names")= chr [1:6] "redirect" "namelookup" "connect" "pretransfer" ...
#>  - attr(*, "Content-Type")= chr "text/csv; charset=utf-8"
#>  - attr(*, "accessed")= POSIXct[1:1], format: "2024-09-19 17:09:20"
str(avs_raw_record)
#>  'rcer_raw_record' chr "record_id,uniform_number,firstname,lastname,hof,nationality,position,birthdate,first_nhl_game,last_nhl_game,hei"| __truncated__
#>  - attr(*, "url")= chr "https://redcap.ucdenver.edu/api/"
#>  - attr(*, "status_code")= int 200
#>  - attr(*, "times")= Named num [1:6] 0.0 1.4e-05 0.0 8.1e-05 3.4e-01 ...
#>   ..- attr(*, "names")= chr [1:6] "redirect" "namelookup" "connect" "pretransfer" ...
#>  - attr(*, "Content-Type")= chr "text/csv; charset=utf-8"
#>  - attr(*, "accessed")= POSIXct[1:1], format: "2024-09-19 17:09:21"
str(avs_raw_core)
#> List of 4
#>  $ project_raw : 'rcer_raw_project' chr "project_id,project_title,creation_time,production_time,in_production,project_language,purpose,purpose_other,pro"| __truncated__
#>   ..- attr(*, "url")= chr "https://redcap.ucdenver.edu/api/"
#>   ..- attr(*, "status_code")= int 200
#>   ..- attr(*, "times")= Named num [1:6] 0 0.000017 0 0.000105 0.142771 ...
#>   .. ..- attr(*, "names")= chr [1:6] "redirect" "namelookup" "connect" "pretransfer" ...
#>   ..- attr(*, "Content-Type")= chr "text/csv; charset=utf-8"
#>   ..- attr(*, "accessed")= POSIXct[1:1], format: "2024-09-19 17:09:21"
#>  $ metadata_raw: 'rcer_raw_metadata' chr "field_name,form_name,section_header,field_type,field_label,select_choices_or_calculations,field_note,text_valid"| __truncated__
#>   ..- attr(*, "url")= chr "https://redcap.ucdenver.edu/api/"
#>   ..- attr(*, "status_code")= int 200
#>   ..- attr(*, "times")= Named num [1:6] 0 0.000017 0 0.000099 0.133926 ...
#>   .. ..- attr(*, "names")= chr [1:6] "redirect" "namelookup" "connect" "pretransfer" ...
#>   ..- attr(*, "Content-Type")= chr "text/csv; charset=utf-8"
#>   ..- attr(*, "accessed")= POSIXct[1:1], format: "2024-09-19 17:09:21"
#>  $ user_raw    : 'rcer_raw_user' chr "username,email,firstname,lastname,expiration,data_access_group,data_access_group_id,design,alerts,user_rights,d"| __truncated__
#>   ..- attr(*, "url")= chr "https://redcap.ucdenver.edu/api/"
#>   ..- attr(*, "status_code")= int 200
#>   ..- attr(*, "times")= Named num [1:6] 0 0.000022 0 0.0001 0.118193 ...
#>   .. ..- attr(*, "names")= chr [1:6] "redirect" "namelookup" "connect" "pretransfer" ...
#>   ..- attr(*, "Content-Type")= chr "text/csv; charset=utf-8"
#>   ..- attr(*, "accessed")= POSIXct[1:1], format: "2024-09-19 17:09:21"
#>  $ record_raw  : 'rcer_raw_record' chr "record_id,uniform_number,firstname,lastname,hof,nationality,position,birthdate,first_nhl_game,last_nhl_game,hei"| __truncated__
#>   ..- attr(*, "url")= chr "https://redcap.ucdenver.edu/api/"
#>   ..- attr(*, "status_code")= int 200
#>   ..- attr(*, "times")= Named num [1:6] 0 0.000015 0 0.0001 0.155303 ...
#>   .. ..- attr(*, "names")= chr [1:6] "redirect" "namelookup" "connect" "pretransfer" ...
#>   ..- attr(*, "Content-Type")= chr "text/csv; charset=utf-8"
#>   ..- attr(*, "accessed")= POSIXct[1:1], format: "2024-09-19 17:09:21"
#>  - attr(*, "class")= chr "rcer_rccore"

avs <- format_record(avs_raw_record, avs_raw_metadata)
str(avs)
#> Classes ‘rcer_record’ and 'data.frame':	32 obs. of  75 variables:
#>  $ record_id                      : chr  "1" "2" "3" "4" ...
#>  $ uniform_number                 : chr  "1" "46" "45" "4" ...
#>  $ firstname                      : chr  "David" "Yuri" "Rick" "Rob" ...
#>  $ lastname                       : chr  "Aebischer" "Babenko" "Berry" "Blake" ...
#>  $ hof                            : int  0 0 0 1 1 0 0 0 0 0 ...
#>  $ nationality                    : chr  "Swiss" "USSR" "Canada" "Canada" ...
#>  $ position                       : Factor w/ 5 levels "Goal","Left Wing",..: 1 4 5 5 5 5 3 2 2 5 ...
#>  $ birthdate                      : Date, format: "1978-02-07" "1978-01-02" ...
#>  $ first_nhl_game                 : Date, format: "2001-04-07" "2000-11-22" ...
#>  $ last_nhl_game                  : Date, format: "2007-10-10" "2000-11-29" ...
#>  $ height                         : int  73 73 74 76 71 74 72 76 70 74 ...
#>  $ weight                         : int  185 200 210 220 219 205 205 235 191 220 ...
#>  $ shoots                         : Factor w/ 2 levels "Left","Right": NA 1 1 2 1 1 2 1 2 2 ...
#>  $ catches                        : Factor w/ 2 levels "Left","Right": 1 NA NA NA NA NA NA NA NA NA ...
#>  $ experience                     : int  0 0 0 11 21 5 6 3 2 9 ...
#>  $ roster_complete                : Factor w/ 3 levels "Incomplete","Unverified",..: 3 3 3 3 3 3 3 3 3 3 ...
#>  $ gp                             : int  26 3 19 13 80 79 39 41 71 35 ...
#>  $ goals                          : int  0 0 0 2 7 5 13 1 24 3 ...
#>  $ assists                        : int  1 0 4 8 52 12 13 1 41 12 ...
#>  $ points                         : num  1 0 4 10 59 17 26 2 65 15 ...
#>  $ plusmn                         : int  0 0 5 11 25 23 -2 -3 6 6 ...
#>  $ pimi                           : int  0 0 38 8 48 51 59 108 47 42 ...
#>  $ goals_ev                       : int  0 0 0 1 3 5 6 1 13 1 ...
#>  $ goals_pp                       : int  0 0 0 1 2 0 7 0 11 1 ...
#>  $ goals_sh                       : int  0 0 0 0 2 0 0 0 0 1 ...
#>  $ goals_gw                       : int  0 0 0 1 0 0 2 0 5 1 ...
#>  $ assists_ev                     : int  NA 0 4 6 21 11 7 1 22 7 ...
#>  $ assists_pp                     : int  NA 0 0 2 31 0 6 0 18 5 ...
#>  $ assists_sh                     : int  NA 0 0 0 0 1 0 0 0 0 ...
#>  $ shots                          : int  0 2 10 44 216 76 86 33 204 59 ...
#>  $ shooting_percentage            : num  0 0 0 4.55 3.24 ...
#>  $ toi                            : int  1393 32 231 339 2088 1351 687 264 1281 888 ...
#>  $ atoi                           :Formal class 'Period' [package "lubridate"] with 6 slots
#>   .. ..@ .Data : num  34 34 8 3 6 6 38 26 3 22 ...
#>   .. ..@ year  : num  0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..@ month : num  0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..@ day   : num  0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..@ hour  : num  0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..@ minute: num  53 10 12 26 26 17 17 6 18 25 ...
#>  $ regular_season_scoring_complete: Factor w/ 3 levels "Incomplete","Unverified",..: 3 3 3 3 3 3 3 3 3 3 ...
#>  $ wins                           : int  12 NA NA NA NA NA NA NA NA NA ...
#>  $ losses                         : int  7 NA NA NA NA NA NA NA NA NA ...
#>  $ ties_otl                       : int  3 NA NA NA NA NA NA NA NA NA ...
#>  $ goals_against                  : int  52 NA NA NA NA NA NA NA NA NA ...
#>  $ shots_against                  : int  538 NA NA NA NA NA NA NA NA NA ...
#>  $ saves                          : int  486 NA NA NA NA NA NA NA NA NA ...
#>  $ save_percentage                : num  0.903 NA NA NA NA ...
#>  $ gaa                            : num  2.24 NA NA NA NA NA NA NA NA NA ...
#>  $ so                             : int  3 NA NA NA NA NA NA NA NA NA ...
#>  $ regular_season_goalies_complete: Factor w/ 3 levels "Incomplete","Unverified",..: 3 3 3 3 3 3 3 3 3 3 ...
#>  $ gp_postseason                  : int  1 0 0 23 21 23 0 16 23 23 ...
#>  $ goals_postseason               : int  0 NA NA 6 4 0 NA 0 11 3 ...
#>  $ assists_postseason             : int  0 NA NA 13 6 1 NA 4 5 4 ...
#>  $ points_postseason              : num  0 NA NA 19 10 1 NA 4 16 7 ...
#>  $ plusmn_postseason              : int  0 NA NA 6 9 5 NA 3 5 5 ...
#>  $ pimi_postseason                : int  0 NA NA 16 12 20 NA 14 4 47 ...
#>  $ goals_ev_postseason            : int  0 NA NA 3 1 0 NA 0 9 2 ...
#>  $ goals_pp_postseason            : int  0 NA NA 3 3 0 NA 0 2 1 ...
#>  $ goals_sh_postseason            : int  0 NA NA 0 0 0 NA 0 0 0 ...
#>  $ goals_gw_postseason            : int  0 NA NA 0 1 0 NA 0 2 1 ...
#>  $ assists_ev_postseason          : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ assists_pp_postseason          : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ assists_sh_postseason          : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ shots_postseason               : int  0 NA NA 83 49 20 NA 8 62 28 ...
#>  $ shooting_percentage_postseason : num  NA NA NA 7.23 8.16 ...
#>  $ toi_postseason                 : int  1 NA NA 677 599 328 NA 101 439 652 ...
#>  $ atoi_postseason                :Formal class 'Period' [package "lubridate"] with 6 slots
#>   .. ..@ .Data : num  32 NA NA 26 32 17 NA 18 6 22 ...
#>   .. ..@ year  : num  0 NA NA 0 0 0 NA 0 0 0 ...
#>   .. ..@ month : num  0 NA NA 0 0 0 NA 0 0 0 ...
#>   .. ..@ day   : num  0 NA NA 0 0 0 NA 0 0 0 ...
#>   .. ..@ hour  : num  0 NA NA 0 0 0 NA 0 0 0 ...
#>   .. ..@ minute: num  0 NA NA 29 28 14 NA 6 19 28 ...
#>  $ post_season_scoring_complete   : Factor w/ 3 levels "Incomplete","Unverified",..: 3 3 3 3 3 3 3 3 3 3 ...
#>  $ wins_postseason                : int  0 NA NA NA NA NA NA NA NA NA ...
#>  $ losses_postseason              : int  0 NA NA NA NA NA NA NA NA NA ...
#>  $ ties_otl_postseason            : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ goals_allowed_postseason       : int  0 NA NA NA NA NA NA NA NA NA ...
#>  $ saves_postseason               : int  0 NA NA NA NA NA NA NA NA NA ...
#>  $ save_percentage_postseason     : num  0 NA NA NA NA NA NA NA NA NA ...
#>  $ gaa_postseason                 : num  0 NA NA NA NA NA NA NA NA NA ...
#>  $ so_postseason                  : int  0 NA NA NA NA NA NA NA NA NA ...
#>  $ post_season_goalies_complete   : Factor w/ 3 levels "Incomplete","Unverified",..: 3 3 3 3 3 3 3 3 3 3 ...
#>  $ eg_checkbox___cb01             : int  1 0 0 0 0 1 1 1 0 1 ...
#>  $ eg_checkbox___cb02             : int  0 1 0 0 1 0 1 1 0 1 ...
#>  $ eg_checkbox___cb03             : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ extras_complete                : Factor w/ 3 levels "Incomplete","Unverified",..: 1 1 1 1 1 1 1 1 1 1 ...